quest_master/requirements.txt
PyQt6>=6.6.0
Jinja2>=3.1.2
weasyprint>=60.0
python-docx>=1.0.0
qrcode[pil]>=7.4.2
Pillow>=10.0.0